requirements.txt
pyln.proto.message
pyln.proto.message.bolt
coincurve
